fixed bug with icons positioning in LandingPage.js, renamed files with styles for Mui
