+++ client/src/app/pages/landing-page/LandingPage.MuiStyles.js
@@ -3,21 +3,11 @@ const styles = {
     height: 'calc(100vh - 64px)',
     paddingBottom: '60px',
   },
-  text: {
+  mainText: {
     backgroundImage: 'linear-gradient(90deg,#12c2e9,#c471ed,#f64f59)',
     '-webkit-background-clip': 'text',
     '-webkit-text-fill-color': 'transparent',
     fontWeight: 'bold',
-  },
-  starLeft: {
-    position: 'absolute',
-    top: '150px',
-    left: '300px',
-  },
-  starRight: {
-    position: 'absolute',
-    top: '150px',
-    right: '300px',
   },
 };
