+++ client/src/app/pages/landing-page/LandingPage.js
@@ -4,7 +4,7 @@ import React from 'react';
 import Header from '../../components/header/Header.js';
 import { ShootingStars } from '../../svg/ShootingStars/ShootingStars.js';
 import { Gear } from '../../svg/Gear/Gear.js';
-import { styles } from './LandingPage.styles.js';
+import { styles } from './LandingPage.MuiStyles.js';
 
 class LandingPage extends React.Component {
   render() {
@@ -13,9 +13,6 @@ class LandingPage extends React.Component {
       <>
         <Header />
         <Container>
-          <Box key={2} className={classes.starLeft}>
-            <ShootingStars />
-          </Box>
           <Grid
             container
             spacing={6}
@@ -24,18 +21,25 @@ class LandingPage extends React.Component {
             direction='column'
             className={classes.mainContainer}
           >
-            <Grid item>
+            <Grid item container justifyContent='space-around'>
+              <Box>
+                <ShootingStars />
+              </Box>
               <Gear />
+              <Box>
+                <ShootingStars />
+              </Box>
             </Grid>
             <Grid item>
-              <Typography variant='h3' align='center' className={classes.text}>
+              <Typography
+                variant='h3'
+                align='center'
+                className={classes.mainText}
+              >
                 THE SUPERIOR FORM BUILDER DESIGNED SPECIFICALLY FOR YOU
               </Typography>
             </Grid>
           </Grid>
-          <Box key={1} className={classes.starRight}>
-            <ShootingStars />
-          </Box>
         </Container>
       </>
     );
